Bulk fix map image urls (#13493)
* Bulk fix triplea_maps.yaml image URLs

(1) The URLs of the form 'github.com..../raw/...' redirect to 'raw.githubusercontent.com'
These URLS are replaced with their redirection targets. Curling these URLs now returns a 200 rather
than a 302. When validating image URLs we can now exclude 302s as not a valid status code.

(2) The URLs of the form 'github.com..../blob/...' pointed to the HTML page containing
the preview image. These URLs return HTML content, not an image! These are corrected.

* Update verify-urls check that images URLs return 200 and not 302 redirects

* Extract github action 'script' in verify-urls...yml to its own script

Having the script portion in its own file:
(1) makes it easier to run, which in turn makes it easier to modify and debug
(2) makes modification easier, no worries about colliding with YAML formatting
and paragraph indentation or special characters

* Update verify-urls script, idiomatic shell usage, favor using '||' over capturing status code and using an if statement

* verify-urls script: simplify, avoid using an intermediate 'txt' file for capturing results, loop over the results directly instead